Split storage into substructs for nodes and pods
